+++ NEWS.md
@@ -1,3 +1,32 @@
+# sandpaper 0.12.3 (2023-06-01)
+
+* A bug where the git credentials are accidentally changed when a lesson is
+  built is fixed by no longer querying git author when the lesson is built.
+  (reported: @joelnitta, @velait, and @zkamvar, #449; fixed: @zkamvar, #476).
+
+# sandpaper 0.12.2 (2023-05-29)
+
+## BUG FIX
+
+* A bug where the sidebar for non-episode pages had extra commas was fixed
+  (reported: @zkamvar, #473; fixed: @zkamvar, #474)
+
+# sandpaper 0.12.1 (2023-05-26)
+
+## BUG FIX
+
+* The current page of the sidebar no longer hides the episode number. 
+  (reported: @cynthiaftw, https://github.com/carpentries/workbench/issues/42 and
+  #432; fixed: @zkamvar, #472)
+- metadata for episodes with titles containing markup no longer include that
+  markup in the metadata (@zkamvar, #472)
+
+## MISC
+
+* The internal function `sandpaper:::check_pandoc()` now points to the correct
+  URL to download RStudio, which moved after the migration to posit (@zkamvar,
+  #471) 
+
 # sandpaper 0.12.0 (2023-05-19)
 
 ## NEW FEATURES